Moving constants from shared_constants.py to shared.py (NaCl.py is now on top of the hierarchy and shared.py is right below but above the subtranspilers and type_processors)
