+++ type_processors/conntrack.py
@@ -17,9 +17,9 @@
 from __future__ import absolute_import
 # To avoid: <...>/NaCl/type_processors/conntrack.py:1: RuntimeWarning: Parent module '<...>/NaCl/type_processors' not found while handling absolute import
 
-from NaCl import NaCl_exception, TCP, UDP, ICMP, Typed, exit_NaCl
+from NaCl import exit_NaCl, NaCl_exception, Typed
 
-from shared import TEMPLATE_KEY_NAME
+from shared import TEMPLATE_KEY_NAME, TCP, UDP, ICMP
 
 # -------------------- CONSTANTS Conntrack --------------------
