+++ type_processors/load_balancer.py
@@ -19,9 +19,11 @@ from __future__ import absolute_import
 
 # TODO later: Own mustache file as well for Load_balancer - then import into main mustache file if possible
 
-from NaCl import Typed, exit_NaCl, predefined_values_cpp, TCP
+from NaCl import exit_NaCl, Typed
 
+# TODO
 from shared import *
+# TCP, predefined_values_cpp
 
 TYPE_LOAD_BALANCER = "load_balancer"
